# openapi.yaml
openapi: 3.1.0
info:
  title: Sediment API
  description: 'Self-hosted ingest: facts in through three doors (ADR 0001)'
  version: 0.1.0
paths:
...
  /v1/me:
    get:
      tags:
      - v1
      summary: Me
      description: 'Auth probe for ``sediment login``: the deployment''s tenant and
        the API

        version, reported verbatim so the client can detect skew.'
      operationId: me_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Me V1 Me Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
      security:
      - bearerAuth: []
      - operatorBearerAuth: []
  /v1/facts:
    get:
      tags:
      - v1
      summary: Facts
      description: 'One entry per ``FactTable``: total rows and the derivation-facing

        ("visible" = not quarantined) count.  Sessions are upserted metadata,

        not quarantinable facts, so they carry a bare count with no visible

        column — the same shape ``sediment facts`` prints.'
      operationId: facts_v1_facts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Facts V1 Facts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '403':
          description: Operator authority required.
      security:
      - operatorBearerAuth: []
  /v1/facts/session/{session_id}:
    get:
      tags:
      - v1
      summary: Session Facts
      description: 'Fact counts for one session, so a client can verify its own writes

        without treating unrelated organization-wide totals as evidence.'
      operationId: session_facts_v1_facts_session__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          pattern: ^[^\u0000\uD800-\uDFFF]*$
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Session Facts V1 Facts Session  Session Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '403':
          description: Operator authority required.
      security:
      - operatorBearerAuth: []
  /v1/facts/session/{session_id}/inference-calls:
    get:
      tags:
      - v1
      summary: Session Inference Calls
      description: 'Return the narrow fields needed to reconcile one Session''s usage.


        Prompts, responses, raw provider payloads, and user identity stay absent.'
      operationId: session_inference_calls_v1_facts_session__session_id__inference_calls_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          pattern: ^[^\u0000\uD800-\uDFFF]*$
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Session Inference Calls V1 Facts Session  Session
                  Id  Inference Calls Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '403':
          description: Operator authority required.
      security:
      - operatorBearerAuth: []
  /v1/facts/session/{session_id}/compatibility-evidence:
    get:
      tags:
      - v1
      summary: Session Compatibility Evidence
      description: Return bounded decision, Edit observation, and inference join fields.
      operationId: session_compatibility_evidence_v1_facts_session__session_id__compatibility_evidence_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          pattern: ^[^\u0000\uD800-\uDFFF]*$
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Session Compatibility Evidence V1 Facts Session  Session
                  Id  Compatibility Evidence Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '403':
          description: Operator authority required.
      security:
      - operatorBearerAuth: []
  /v1/reports/model-outcomes:
    get:
      tags:
      - reports
      summary: Model Outcomes Report
      description: Return a bounded model-outcome report.
      operationId: model_outcomes_report_v1_reports_model_outcomes_get
      parameters:
      - name: cohort_start
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Cohort Start
      - name: cohort_end
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Cohort End
      - name: as_of
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: As Of
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportEnvelope'
        '503':
          description: Database unavailable, or work capacity or execution budget
            exceeded.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '403':
          description: Operator authority required.
      security:
      - operatorBearerAuth: []
  /v1/reports/accepted-work-lifecycle:
    get:
      tags:
      - reports
      summary: Accepted Work Lifecycle Report
      description: Return a bounded accepted-work lifecycle report.
      operationId: accepted_work_lifecycle_report_v1_reports_accepted_work_lifecycle_get
      parameters:
      - name: cohort_start
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Cohort Start
      - name: cohort_end
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Cohort End
      - name: as_of
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: As Of
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportEnvelope'
        '503':
          description: Database unavailable, or work capacity or execution budget
            exceeded.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '403':
          description: Operator authority required.
      security:
      - operatorBearerAuth: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReportEnvelope:
      properties:
        schema_version:
          type: integer
          const: 1
          title: Schema Version
          default: 1
        scope:
          $ref: '#/components/schemas/ReportScopeResponse'
        report:
          additionalProperties: true
          type: object
          title: Report
      type: object
      required:
      - scope
      - report
      title: ReportEnvelope
      description: Version 1 operational-report response.
    ReportScopeResponse:
      properties:
        cohort_start:
          type: string
          format: date-time
          title: Cohort Start
        cohort_end:
          type: string
          format: date-time
          title: Cohort End
        as_of:
          type: string
          format: date-time
          title: As Of
        max_inference_calls:
          type: integer
          const: 50000
          title: Max Inference Calls
      type: object
      required:
      - cohort_start
      - cohort_end
      - as_of
      - max_inference_calls
      title: ReportScopeResponse
      description: The explicit bounds applied to one operational report.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Named SEDIMENT_INGEST_TOKENS credential or legacy SEDIMENT_API_BEARER_TOKEN.
        Capture authority only; missing/invalid returns401.
    operatorBearerAuth:
      type: http
      scheme: bearer
      description: SEDIMENT_OPERATOR_TOKEN. Operator access and explicit operator
        ingest; missing/invalid returns401, ingest authority on an operator route
        returns403.
    githubWebhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: GitHub webhook HMAC-SHA256 of the raw body, keyed with SEDIMENT_GITHUB_WEBHOOK_SECRET.
        Missing/invalid -> 401.